restructuring the folders and files to a better go standard
