+++ callouts/go/extproc/config/Dockerfile
@@ -19,13 +19,13 @@ FROM golang:1.18-alpine AS builder
 WORKDIR /app
 
 # Copy go mod and sum files
-COPY go.mod go.sum ./
+COPY ../../go.mod go.sum ./
 
 # Download all dependencies. Dependencies will be cached if the go.mod and go.sum files are not changed
 RUN go mod download
 
 # Copy the source from the current directory to the Working Directory inside the container
-COPY . .
+COPY ../.. .
 
 # Build the Go app
 ARG EXAMPLE_TYPE
