+++ callouts/go/extproc/config/docker-compose.yml
@@ -17,7 +17,7 @@ version: '3.8'
 services:
   redirect:
     build:
-      context: .
+      context: ../..
       dockerfile: Dockerfile
       args:
         - EXAMPLE_TYPE=redirect
