+++ callouts/go/extproc/pkg/utils/callout_tools.go
